Minor renaming fixes
