+++ packages/twenty-front/src/modules/object-record/record-table/components/RecordTable.tsx
@@ -1,154 +1,111 @@
-import { useRef } from 'react';
+import { useContext } from 'react';
 import styled from '@emotion/styled';
-import { useRecoilCallback, useRecoilValue } from 'recoil';
 
-import { useObjectMetadataItem } from '@/object-metadata/hooks/useObjectMetadataItem';
-import { useObjectNameSingularFromPlural } from '@/object-metadata/hooks/useObjectNameSingularFromPlural';
-import { ObjectsTable } from '@/object-record/record-table/components/ObjectsTable';
-import { ScrollLeftEdge } from '@/object-record/record-table/components/ScrollLeftEdge';
-import { isRecordTableInitialLoadingState } from '@/object-record/record-table/states/isRecordTableInitialLoadingState';
-import { IconPlus } from '@/ui/display/icon';
-import { Button } from '@/ui/input/button/components/Button';
-import { DragSelect } from '@/ui/utilities/drag-select/components/DragSelect';
-import { ObjectsTableWrapper } from '@/ui/utilities/objects-table/components/ObjectsTableWrapper';
-import { ScrollWrapper } from '@/ui/utilities/scroll/components/ScrollWrapper';
-import { useViewFields } from '@/views/hooks/internal/useViewFields';
-import { mapColumnDefinitionsToViewFields } from '@/views/utils/mapColumnDefinitionToViewField';
+import { RecordTableBody } from '@/object-record/record-table/components/RecordTableBody';
+import { RecordTableBodyEffect } from '@/object-record/record-table/components/RecordTableBodyEffect';
+import { RecordTableHeader } from '@/object-record/record-table/components/RecordTableHeader';
+import { RecordTableRefContext } from '@/object-record/record-table/contexts/RecordTableRefContext';
+import { rgba } from '@/ui/theme/constants/colors';
 
-import { RecordUpdateContext } from '../contexts/EntityUpdateMutationHookContext';
-import { useRecordTable } from '../hooks/useRecordTable';
-import { RecordTableScope } from '../scopes/RecordTableScope';
-import { numberOfTableRowsState } from '../states/numberOfTableRowsState';
+const StyledTable = styled.table`
+  border-radius: ${({ theme }) => theme.border.radius.sm};
+  border-spacing: 0;
+  margin-right: ${({ theme }) => theme.table.horizontalCellMargin};
+  table-layout: fixed;
 
-import { RecordTableInternalEffect } from './RecordTableInternalEffect';
+  width: calc(100% - ${({ theme }) => theme.table.horizontalCellMargin} * 2);
 
-const StyledObjectEmptyContainer = styled.div`
-  align-items: center;
-  align-self: stretch;
-  display: flex;
-  flex: 1 0 0;
-  flex-direction: column;
-  gap: ${({ theme }) => theme.spacing(2)};
-  justify-content: center;
-  padding-bottom: ${({ theme }) => theme.spacing(16)};
-  padding-left: ${({ theme }) => theme.spacing(4)};
-  padding-right: ${({ theme }) => theme.spacing(4)};
-  padding-top: ${({ theme }) => theme.spacing(3)};
-`;
+  th {
+    border-block: 1px solid ${({ theme }) => theme.border.color.light};
+    color: ${({ theme }) => theme.font.color.tertiary};
+    padding: 0;
+    text-align: left;
 
-const StyledEmptyObjectTitle = styled.div`
-  color: ${({ theme }) => theme.font.color.secondary};
-  font-size: ${({ theme }) => theme.font.size.xxl};
-  font-weight: ${({ theme }) => theme.font.weight.semiBold};
-  line-height: ${({ theme }) => theme.text.lineHeight.md};
-`;
+    :last-child {
+      border-right-color: transparent;
+    }
+    :first-of-type {
+      border-left-color: transparent;
+      border-right-color: transparent;
+    }
+  }
 
-const StyledEmptyObjectSubTitle = styled.div`
-  color: ${({ theme }) => theme.font.color.extraLight};
-  font-size: ${({ theme }) => theme.font.size.xxl};
-  font-weight: ${({ theme }) => theme.font.weight.semiBold};
-  line-height: ${({ theme }) => theme.text.lineHeight.md};
-  margin-bottom: ${({ theme }) => theme.spacing(2)};
-`;
+  td {
+    border-bottom: 1px solid ${({ theme }) => theme.border.color.light};
+    color: ${({ theme }) => theme.font.color.primary};
+    padding: 0;
 
-const StyledTableWithHeader = styled.div`
-  display: flex;
-  flex: 1;
-  flex-direction: column;
-  width: 100%;
-`;
+    text-align: left;
 
-const StyledTableContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-  position: relative;
+    :last-child {
+      border-right-color: transparent;
+    }
+    :first-of-type {
+      border-left-color: transparent;
+      border-right-color: transparent;
+    }
+  }
+
+  th,
+  td {
+    background-color: ${({ theme }) => theme.background.primary};
+    border-right: 1px solid ${({ theme }) => theme.border.color.light};
+  }
+
+  thead th:nth-of-type(-n + 2),
+  tbody td:nth-of-type(-n + 2) {
+    position: sticky;
+    z-index: 2;
+    border-right: none;
+  }
+
+  thead th:nth-of-type(1),
+  tbody td:nth-of-type(1) {
+    left: 0;
+  }
+  thead th:nth-of-type(2),
+  tbody td:nth-of-type(2) {
+    left: calc(${({ theme }) => theme.table.checkboxColumnWidth} - 2px);
+  }
+
+  tbody td:nth-of-type(2)::after,
+  thead th:nth-of-type(2)::after {
+    content: '';
+    height: calc(100% + 1px);
+    position: absolute;
+    top: 0;
+    width: 4px;
+    right: -4px;
+  }
+
+  &.freeze-first-columns-shadow thead th:nth-of-type(2)::after,
+  &.freeze-first-columns-shadow tbody td:nth-of-type(2)::after {
+    box-shadow: ${({ theme }) =>
+      `4px 0px 4px -4px ${
+        theme.name === 'dark'
+          ? rgba(theme.grayScale.gray50, 0.8)
+          : rgba(theme.grayScale.gray100, 0.25)
+      } inset`};
+  }
+
+  thead th:nth-of-type(3),
+  tbody td:nth-of-type(3) {
+    border-left: 1px solid ${({ theme }) => theme.border.color.light};
+  }
 `;
 
 type RecordTableProps = {
-  recordTableId: string;
-  viewBarId: string;
-  updateRecordMutation: (params: any) => void;
   createRecord: () => void;
 };
 
-export const RecordTable = ({
-  updateRecordMutation,
-  createRecord,
-  recordTableId,
-  viewBarId,
-}: RecordTableProps) => {
-  const tableBodyRef = useRef<HTMLDivElement>(null);
-
-  const numberOfTableRows = useRecoilValue(numberOfTableRowsState);
-
-  const isRecordTableInitialLoading = useRecoilValue(
-    isRecordTableInitialLoadingState,
-  );
-
-  const {
-    scopeId: objectNamePlural,
-    resetTableRowSelection,
-    setRowSelectedState,
-  } = useRecordTable({
-    recordTableScopeId: recordTableId,
-  });
-
-  const { objectNameSingular } = useObjectNameSingularFromPlural({
-    objectNamePlural,
-  });
-
-  const { objectMetadataItem: foundObjectMetadataItem } = useObjectMetadataItem(
-    {
-      objectNameSingular,
-    },
-  );
-
-  const { persistViewFields } = useViewFields(viewBarId);
+export const RecordTable = ({ createRecord }: RecordTableProps) => {
+  const recordTableRef = useContext(RecordTableRefContext);
 
   return (
-    <RecordTableScope
-      recordTableScopeId={recordTableId}
-      onColumnsChange={useRecoilCallback(() => (columns) => {
-        persistViewFields(mapColumnDefinitionsToViewFields(columns));
-      })}
-    >
-      <ScrollWrapper>
-        <ObjectsTableWrapper>
-          <ScrollLeftEdge />
-          <RecordUpdateContext.Provider value={updateRecordMutation}>
-            <StyledTableWithHeader>
-              <StyledTableContainer>
-                <div ref={tableBodyRef}>
-                  <ObjectsTable createRecord={createRecord} />
-                  <DragSelect
-                    dragSelectable={tableBodyRef}
-                    onDragSelectionStart={resetTableRowSelection}
-                    onDragSelectionChange={setRowSelectedState}
-                  />
-                </div>
-                <RecordTableInternalEffect tableBodyRef={tableBodyRef} />
-                {!isRecordTableInitialLoading && numberOfTableRows === 0 && (
-                  <StyledObjectEmptyContainer>
-                    <StyledEmptyObjectTitle>
-                      No {foundObjectMetadataItem?.namePlural}
-                    </StyledEmptyObjectTitle>
-                    <StyledEmptyObjectSubTitle>
-                      Create one:
-                    </StyledEmptyObjectSubTitle>
-                    <Button
-                      Icon={IconPlus}
-                      title={`Add a ${foundObjectMetadataItem?.nameSingular}`}
-                      variant={'secondary'}
-                      onClick={createRecord}
-                    />
-                  </StyledObjectEmptyContainer>
-                )}
-              </StyledTableContainer>
-            </StyledTableWithHeader>
-          </RecordUpdateContext.Provider>
-        </ObjectsTableWrapper>
-      </ScrollWrapper>
-    </RecordTableScope>
+    <StyledTable ref={recordTableRef} className="entity-table-cell">
+      <RecordTableHeader createRecord={createRecord} />
+      <RecordTableBodyEffect />
+      <RecordTableBody />
+    </StyledTable>
   );
 };
