+++ packages/twenty-front/src/modules/object-record/record-table/components/RecordTableFirstColumnScrollObserver.tsx
@@ -1,17 +1,17 @@
 import { useContext } from 'react';
 import { useInView } from 'react-intersection-observer';
 
-import { ObjectsTableWrapperContext } from '@/ui/utilities/objects-table/components/ObjectsTableWrapper';
+import { RecordTableRefContext } from '@/object-record/record-table/contexts/RecordTableRefContext';
 import { ScrollWrapperContext } from '@/ui/utilities/scroll/components/ScrollWrapper';
 
-export const ScrollLeftEdge = () => {
+export const RecordTableFirstColumnScrollObserver = () => {
   const scrollWrapperRef = useContext(ScrollWrapperContext);
-  const objectsTableWrapperRef = useContext(ObjectsTableWrapperContext);
+  const recordTableRef = useContext(RecordTableRefContext);
 
   const { ref: elementRef } = useInView({
     root: scrollWrapperRef.current,
     onChange: (inView) => {
-      objectsTableWrapperRef.current?.classList.toggle(
+      recordTableRef.current?.classList.toggle(
         'freeze-first-columns-shadow',
         !inView,
       );
