+++ packages/twenty-front/src/modules/object-record/record-table/contexts/RecordTableRefContext.ts
@@ -1,0 +1,7 @@
+import { createContext, RefObject } from 'react';
+
+export const RecordTableRefContext = createContext<RefObject<HTMLTableElement>>(
+  {
+    current: null,
+  },
+);
